fix: applyCSSMinimizer
