+++ packages/builder/builder-shared/package.json
@@ -155,6 +155,7 @@
     "acorn": "^8.10.0",
     "caniuse-lite": "^1.0.30001520",
     "cssnano": "6.0.1",
+    "css-minimizer-webpack-plugin": "5.0.1",
     "htmlparser2": "9.0.0",
     "line-diff": "2.1.1",
     "postcss": "8.4.27",
