+++ packages/builder/builder-webpack-provider/package.json
@@ -100,7 +100,6 @@
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.10",
     "@swc/helpers": "0.5.1",
     "caniuse-lite": "^1.0.30001520",
-    "css-minimizer-webpack-plugin": "5.0.1",
     "html-webpack-plugin": "5.5.3",
     "mini-css-extract-plugin": "2.7.6",
     "postcss": "8.4.27",
