+++ packages/builder/builder-webpack-provider/src/plugins/minimize.ts
@@ -1,16 +1,12 @@
 import { CHAIN_ID } from '@modern-js/utils/chain-id';
 import {
-  getCssnanoDefaultOptions,
+  applyCSSMinimizer,
+  BundlerChain,
   getJSMinifyOptions,
 } from '@modern-js/builder-shared';
-import type {
-  WebpackChain,
-  BuilderPlugin,
-  CssMinimizerPluginOptions,
-  NormalizedConfig,
-} from '../types';
+import type { BuilderPlugin, NormalizedConfig } from '../types';
 
-async function applyJSMinimizer(chain: WebpackChain, config: NormalizedConfig) {
+async function applyJSMinimizer(chain: BundlerChain, config: NormalizedConfig) {
   const { default: TerserPlugin } = await import('terser-webpack-plugin');
 
   const finalOptions = await getJSMinifyOptions(config);
@@ -25,37 +21,11 @@ async function applyJSMinimizer(chain: WebpackChain, config: NormalizedConfig) {
     .end();
 }
 
-async function applyCSSMinimizer(
-  chain: WebpackChain,
-  config: NormalizedConfig,
-) {
-  const { CHAIN_ID, applyOptionsChain } = await import('@modern-js/utils');
-  const { default: CssMinimizerPlugin } = await import(
-    'css-minimizer-webpack-plugin'
-  );
-
-  const mergedOptions: CssMinimizerPluginOptions = applyOptionsChain(
-    {
-      minimizerOptions: getCssnanoDefaultOptions(),
-    },
-    config.tools.minifyCss,
-  );
-
-  chain.optimization
-    .minimizer(CHAIN_ID.MINIMIZER.CSS)
-    .use(CssMinimizerPlugin, [
-      // Due to css-minimizer-webpack-plugin has changed the type of class, which using a generic type in
-      // constructor, leading auto inference of parameters of plugin constructor is not possible, using any instead
-      mergedOptions as any,
-    ])
-    .end();
-}
-
 export const builderPluginMinimize = (): BuilderPlugin => ({
   name: 'builder-plugin-minimize',
 
   setup(api) {
-    api.modifyWebpackChain(async (chain, { isProd }) => {
+    api.modifyBundlerChain(async (chain, { isProd }) => {
       const config = api.getNormalizedConfig();
       const isMinimize = isProd && !config.output.disableMinimize;
